test: Access Chicago data file for DataFrame creation
As a first step, this code is written to check and verify that the
chicigo.csv is loaded correctly and that the accompanying DataFrame is
created as a result.

One issue to be looked at further is that I experienced problems in
accessing the csv file when it is in a location outside of the directory
that the test_script.py file is in.