fix: Correct error for Practice Problem 2 - Days Name
As I was working through Step 4 of Practice Problem 3, I realised that I
had misread the instructions for Practice Problem 2, specifically the
part regarding the output of the Days of the Week. Originally, and
mistakenly, my output was an integer, representing the day of the week.

My task was obtaining the output in the form of the name of the day.
This commit fixes this issue throughout the code, and I can now move on
to solving Step 4 of Practice Problem 3.